fix(vault-provider): moved sidebar to vault files
