+++ src/routes/index.tsx
@@ -7,8 +7,9 @@ import { dappRoutes } from '@/modules/dapp';
 // import { exampleRoutes } from '@/modules/example';
 import { homeRoutes } from '@/modules/home';
 import { transactionRoutes } from '@/modules/transactions';
-import { VaultRoutes } from '@/modules/vault';
+import { vaultRoutes } from '@/modules/vault';
 import { workspaceRoutes } from '@/modules/workspace';
+import { VaultInfosProvider } from '@/modules/vault/providers/VaultInfosProvider';
 // import { VaultInfosProvider } from '@/modules/vault/providers/VaultInfosProvider';
 
 const routes = (
@@ -17,7 +18,7 @@ const routes = (
     {signinRoutes}
     {homeRoutes}
     {dappRoutes}
-    {VaultRoutes}
+    {vaultRoutes}
     {transactionRoutes}
     {addressBookRoutes}
     {workspaceRoutes}
@@ -27,8 +28,10 @@ const routes = (
 const AppRoutes = () => {
   return (
     <BrowserRouter>
-      <ScrollTop />
-      <Routes>{routes}</Routes>
+      <VaultInfosProvider>
+        <ScrollTop />
+        <Routes>{routes}</Routes>
+      </VaultInfosProvider>
     </BrowserRouter>
   );
 };
